Add files via upload
Library for controlling the motors of the drone. It also includes a code for finding the offsets of each motor in case that we need to run a calibration.

The library uses the function set_micro_scaled() for sending the desired speed of the motors as the width of the PWM pulse in microseconds. The resolution of the pulses is 4 microseconds, therefore, with the ESC calibrated from 1000 to 2000 microseconds we can control the speed from 0 to 250.