Add files via upload
Final code after fixing more things from the original final code. I created a new branch because we didn't checked if all the files that I have are correct for the PI part.

TO DO: (In case that we continue with this at some point)

I don't think that the code is able to create again the same cool plots that we used for the poster and the report, but maybe it does hehe.
Also, we forgot to "calibrate" the vision algorithm. As we do with the barometer for example, removing the offset measured for the initial position. That's also why the plots show how the vision algorithm converged near other values instead of 0, because we never calibrated it.

After calibrating the vision algorithm, I expect the ekf to work much better, since the part of the IMU is calibrated to 0.